rename result pkg to results
