+++ operator/pkg/controllers/controller.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/awslabs/kit/operator/pkg/apis/infrastructure/v1alpha1"
 	"github.com/awslabs/kit/operator/pkg/errors"
-	"github.com/awslabs/kit/operator/pkg/result"
+	"github.com/awslabs/kit/operator/pkg/results"
 	"go.uber.org/zap"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/sets"
@@ -48,7 +48,7 @@ func (c *GenericController) Reconcile(ctx context.Context, req reconcile.Request
 		if errors.IsNotFound(err) {
 			return reconcile.Result{}, nil
 		}
-		return *result.Failed, err
+		return *results.Failed, err
 	}
 	// 2. Copy object for merge patch base
 	persisted := resource.DeepCopyObject()
@@ -56,13 +56,13 @@ func (c *GenericController) Reconcile(ctx context.Context, req reconcile.Request
 	reconcileResult, reconcileErr := c.reconcile(ctx, resource, persisted)
 	// 4. Update Status using a merge patch, we want to set status even when reconcile errored
 	if err := c.Status().Patch(ctx, resource, client.MergeFrom(persisted)); err != nil && !errors.IsNotFound(err) {
-		return *result.Failed, fmt.Errorf("status patch for %s, %w,", req.NamespacedName, err)
+		return *results.Failed, fmt.Errorf("status patch for %s, %w,", req.NamespacedName, err)
 	}
 	if reconcileErr != nil {
 		if errors.IsWaitingForSubResource(reconcileErr) {
-			return *result.Waiting, nil
+			return *results.Waiting, nil
 		}
-		return *result.Failed, reconcileErr
+		return *results.Failed, reconcileErr
 	}
 	return reconcileResult, nil
 }
@@ -79,12 +79,12 @@ func (c *GenericController) reconcile(ctx context.Context, resource Object, pers
 		reconcileResult, err = c.Controller.Reconcile(ctx, resource)
 		if err != nil {
 			resource.StatusConditions().MarkFalse(v1alpha1.Active, "", err.Error())
-			return *result.Failed, fmt.Errorf("reconciling resource, %w", err)
+			return *results.Failed, fmt.Errorf("reconciling resource, %w", err)
 		}
 		resource.StatusConditions().MarkTrue(v1alpha1.Active)
 	} else {
 		if reconcileResult, err = c.Controller.Finalize(ctx, resource); err != nil {
-			return *result.Failed, fmt.Errorf("finalizing resource controller %v, %w", c.Controller.Name(), err)
+			return *results.Failed, fmt.Errorf("finalizing resource controller %v, %w", c.Controller.Name(), err)
 		}
 		// Remove finalizer for this controller
 		resource.SetFinalizers(existingFinalizerSet.Difference(finalizerStr).UnsortedList())
@@ -93,7 +93,7 @@ func (c *GenericController) reconcile(ctx context.Context, resource Object, pers
 	// If the finalizers have changed merge patch the object
 	if !reflect.DeepEqual(existingFinalizers, resource.GetFinalizers()) {
 		if err := c.Patch(ctx, resource, client.MergeFrom(persisted)); err != nil {
-			return *result.Failed, fmt.Errorf("patch object %s, %w", resource.GetName(), err)
+			return *results.Failed, fmt.Errorf("patch object %s, %w", resource.GetName(), err)
 		}
 	}
 	return *reconcileResult, nil
