+++ operator/pkg/controllers/controlplane/controlplane.go
@@ -24,7 +24,7 @@ import (
 	"github.com/awslabs/kit/operator/pkg/controllers/etcd"
 	"github.com/awslabs/kit/operator/pkg/controllers/master"
 	"github.com/awslabs/kit/operator/pkg/kubeprovider"
-	"github.com/awslabs/kit/operator/pkg/result"
+	"github.com/awslabs/kit/operator/pkg/results"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
@@ -53,22 +53,23 @@ func (c *controlPlane) For() controllers.Object {
 	return &v1alpha1.ControlPlane{}
 }
 
+type reconciler func(ctx context.Context, controlPlane *v1alpha1.ControlPlane) (err error)
+
 // Reconcile will check if the resource exists is AWS if it does sync status,
 // else create the resource and then sync status with the ControlPlane.Status
 // object
 func (c *controlPlane) Reconcile(ctx context.Context, object controllers.Object) (res *reconcile.Result, err error) {
-	controlPlane := object.(*v1alpha1.ControlPlane)
-	// TODO create karpenter provisioner spec for creating new nodes.
-	// deploy etcd to the management cluster
-	if err := c.etcdController.Reconcile(ctx, controlPlane); err != nil {
-		return nil, fmt.Errorf("reconciling etcd for cluster %v, %w", controlPlane.ClusterName(), err)
+	for _, reconcile := range []reconciler{
+		c.etcdController.Reconcile,
+		c.masterController.Reconcile,
+	} {
+		if err := reconcile(ctx, object.(*v1alpha1.ControlPlane)); err != nil {
+			return nil, fmt.Errorf("reconciling, %w", err)
+		}
 	}
-	if err := c.masterController.Reconcile(ctx, controlPlane); err != nil {
-		return nil, fmt.Errorf("reconciling master for cluster %v, %w", controlPlane.ClusterName(), err)
-	}
-	return result.Created, nil
+	return results.Created, nil
 }
 
 func (c *controlPlane) Finalize(_ context.Context, _ controllers.Object) (*reconcile.Result, error) {
-	return result.Terminated, nil
+	return results.Terminated, nil
 }
